// frontend/src/app/core/api/model/company-response.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CountryResponse } from './country-response';
import { IndustryResponse } from './industry-response';


export interface CompanyResponse { 
    id?: string;
    name?: string;
    country?: CountryResponse;
    industry?: IndustryResponse;
}