// frontend/src/app/core/api/model/company-add-request.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CompanyAddRequest { 
    name: string;
    countryCode: string;
    industryId: string;
}